:wrench:fix: separate scroll event to custom hook
