+++ src/app/(main)/best/page.tsx
@@ -1,5 +1,3 @@
-import React from 'react'
-
 export default function Best() {
   return <div>Best</div>
 }
