+++ src/components/layout/MainHeader.tsx
@@ -5,30 +5,10 @@ import BasketIcon from '../icons/auth-service-header/BasketIcon'
 import Image from 'next/image'
 import MainHeaderLogo from '../icons/main-header/MainHeaderLogo.png'
 import Link from 'next/link'
-import { useEffect, useState } from 'react'
+import { useScrollEvent } from '@/hooks/UseScrollEvent'
 
 export default function MainHeader() {
-  const [isVisible, setIsVisible] = useState<boolean>(false)
-  const [lastScrollY, setLastScrollY] = useState<number>(0)
-
-  const handleScroll = () => {
-    const currentScrollY = window.scrollY
-    if (currentScrollY > lastScrollY) {
-      // 스크롤내리기
-
-      setIsVisible(false)
-    } else if (currentScrollY < lastScrollY) {
-      // 스크롤올리기
-      setIsVisible(true)
-    }
-    setLastScrollY(currentScrollY)
-  }
-  useEffect(() => {
-    window.addEventListener('scroll', handleScroll)
-    handleScroll()
-
-    return () => window.removeEventListener('scroll', handleScroll)
-  }, [lastScrollY])
+  const isVisible = useScrollEvent()
 
   return (
     <header
