+++ src/components/sign-up/simple/esseitial-form-item/EmailInput.tsx
@@ -2,8 +2,7 @@ import InputWithClear from '@/components/dummy/InputWithClear'
 import React, { useState } from 'react'
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select'
 import { emailDomainData } from '@/datas/dummy/sign-up/SignUpDatas'
-import { useFormValidation } from '../hooks/UseFormValidation'
-
+import { useFormValidation } from '@/hooks/UseFormValidation'
 export default function EmailInput() {
   const { values, isEmpty, errors, setValues, setErrors, handleChange, handleBlur } = useFormValidation()
   const [isEditable, setIsEditable] = useState(true)
